Feat/categories

+++ src/repositories/userRepository.ts
@@ -1,4 +1,4 @@
-import { User } from '@prisma/client';
+import { User, Participants } from '@prisma/client';
 import { prisma } from '../../prisma/client';
 
 class UserRepository {
@@ -16,10 +16,35 @@ class UserRepository {
     }
 
     public async deleteUser(id: number): Promise<User> {
-        return await prisma.user.delete({
-            where: { id }
+        const user = await prisma.user.findUnique({
+          where: { id },
+          include: { participants: true },
         });
-    }
+        
+        if (user) {
+          const ownedGroups = await prisma.group.findMany({
+            where: { ownerEmail: user.email },
+          })
+
+          await prisma.group.deleteMany({
+            where: { id: { 
+              in: ownedGroups.map(group => group.id)
+             } 
+            }
+          })
+
+          await prisma.participants.deleteMany({
+            where: { userEmail: user.email },
+          });
+      
+          return await prisma.user.delete({
+            where: { id },
+          });
+        } else {
+          throw new Error("Usuário não encontrado");
+        }
+      }
+      
 
     public async findAllUsers() {
 
@@ -37,6 +62,26 @@ class UserRepository {
         return await prisma.user.findUnique({
             where: {email}
         })
+    }
+
+    async findUsersByName(name: string) {
+      return await prisma.user.findMany({
+        where: {
+          name: {
+            contains: name,
+            mode: 'insensitive',
+          },
+        },
+        select: {
+          id: true,
+          name: true,
+          email: true,
+        },
+        take: 5, // Limita para evitar sobrecarga
+        orderBy: {
+          name: 'asc',
+        },
+      });
     }
 }
